moved trunc from fitting
